simple_ai_agents/blog_writing_agent/pyproject.toml
[project]
name = "ai-blog-writing-agent"
version = "0.1.0"
description = "AI-powered blog writing tool that analyzes your writing style and generates new content"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.48.1",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "pypdf>=5.1.0",
    "memorisdk>=1.0.1",
    "python-docx>=1.2.0",
]
